ft-#68: Crete new test file to fix all glitches in tests
